Support use_actual_email_prepended_to_subject
In the source, there was an option use_actual_email_prepended_to_subject
defined but not implemented. This commit enables support for that option
(only changing the subject line when there are "real recipients" and the
option is turned on.

The option is off by default, so this would change the default
functionality of people who upgrade. This commit is missing a test and
CHANGELOG entry.